Suggestion Servlet has bugs
