+++ pom.xml
@@ -30,6 +30,12 @@
       <version>1.9.59</version>
     </dependency>
 
+    <dependency>
+      <groupId>com.google.code.gson</groupId>
+      <artifactId>gson</artifactId>
+      <version>2.8.6</version>
+    </dependency>
+
   </dependencies>
 
   <build>
