+++ src/main/java/com/google/sps/data/Suggestion.java
@@ -1,0 +1,17 @@
+package com.google.sps.data;
+
+/** An item for a suggestion */
+public final class Suggestion {
+
+  private final long id;
+  private final String platform;
+  private final String addition;
+  private final long timestamp;
+
+  public Suggestion(long id, String platform, String addition, long timestamp) {
+    this.id = id;
+    this.platform = platform;
+    this.addition = addition;
+    this.timestamp = timestamp;
+  }
+}
