+++ src/main/java/com/google/sps/servlets/SuggestionServlet.java
@@ -3,6 +3,7 @@ package com.google.sps.servlets;
 import com.google.appengine.api.datastore.DatastoreService;
 import com.google.appengine.api.datastore.DatastoreServiceFactory;
 import com.google.appengine.api.datastore.Entity;
+import com.google.sps.data.Suggestion;
 import java.io.IOException;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
@@ -10,7 +11,7 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 /** Servlet responsible for creating a suggestion. */
-@WebServlet("/suggestion")
+@WebServlet("/new-suggestion")
 public class SuggestionServlet extends HttpServlet {
 
   @Override
@@ -27,6 +28,33 @@ public class SuggestionServlet extends HttpServlet {
     DatastoreService datastore = DatastoreServiceFactory.getDatastoreService();
     datastore.put(suggestionEntity);
 
-    response.sendRedirect("/tutorials.html");
+    response.sendRedirect("/index.html");
   }
+
+  /*
+  // LIST suggestions
+  public void doGet(HttpServletRequest request, HttpServletResponse response) throws IOException {
+    Query query = new Query("Task").addSort("timestamp", SortDirection.DESCENDING);
+    
+    DatastoreService datastore = DatastoreServiceFactory.getDatastoreService();
+    PreparedQuery results = datastore.prepare(query);
+    List<Suggestion> suggestions = new ArrayList<>();                  
+    
+    for (Entity entity : results.asIterable()) {
+      long id = entity.getKey().getId();
+      String platform = (String) entity.getProperty("platform");
+      String addition = (String) entity.getProperty("addition");
+      long timestamp = (long) entity.getProperty("timestamp");
+     
+      Suggestion suggestion = new Suggestion(id, platform, addition, timestamp);
+      suggestions.add(suggestion);
+    }
+    
+    Gson gson = new Gson();
+    
+    response.setContentType("application/json;");
+    response.getWriter().println(gson.toJson(suggestions));
+  }
+  */
+
 }
