+++ src/main/webapp/index.html
@@ -1,0 +1,66 @@
+<!DOCTYPE html>
+<html>
+    <head>
+        <meta charset="UTF-8">
+        <title>Seniors Connect</title>
+        <link rel="stylesheet" href="style.css">
+        <script src="script.js"></script>
+    </head>
+    <body>
+        <div class="navbar">
+            <ul>
+                 <!-- TODO: Logo w/ name in upper left corner and nav bar in right <li><a href="index.html">Seniors Connect</a></li>-->
+                 <li><a href="index.html">Home</a></li>
+                 <li><a href="techassistant.html">Tech Assistant</a></li>
+                 <li><a id="login" href="login.html">Login</a></li>
+                 <!-- <li><a id="logout">Logout</a></li>-->
+            </ul>
+        </div>
+        <div class="content">
+            <div class="header">
+                <h1>Tutorials</h1>
+                <p>On this page you will find a list of tools that we have tutorials available for. Click a tutorial to learn more.</p>
+                <p> Watch the video to learn how to use this page </p>
+                <div id="tutorials-video">
+                    <iframe width="560" height="315" src="https://www.youtube.com/embed/BDliEq_0qeQ" frameborder="0" allow="accelerometer; autoplay; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe>
+                </div>
+            </div>
+            
+            <div id = "tutorials">
+                <div id = "left-tutorials-list">
+                    <h2> Currently Available Tutorials </h2>
+                    <li><a href="zoom.html">Zoom</a></li>
+                    <li><a href="https://www.facebook.com/">Facebook</a></li>
+                    <li><a href="https://meet.google.com/">Google Meet</a></li>
+                </div>
+            
+                <div id="right-tutorials-list">
+                    <div id="new-suggestion">
+                        <h2>Create a New Suggestion</h2>
+                        <form method="POST" action="/new-suggestion" class="suggestion-form"></form>
+                            <label for="platform">Platform (ex: Zoom, Facebook etc.):</label><br>
+                            <input type="text" id= "platform" name="platform"><br>
+                            <label for="addition">Additional Details & Specific Questions:</label><br>
+                            <input type="text" id = "addition" name="addition"><br>
+                            <input type="submit" value="Submit">
+                        </form>
+                    </div>
+
+                    <div id="new-suggestion-logged-out">
+                        <h2>Can't find what you are looking for? Make a suggestion.</h2>
+                        <form action="login.html">
+                            <input type="submit" value="Login" />
+                        </form>
+                        <h2>Don't have an Account?</h2>
+                        <form action="signup.html">
+                            <input type="submit" value="Sign Up" />
+                        </form>
+                    </div>
+                </div>
+            </div>
+
+
+        </div>
+    </body>
+</html>
+
